+++ src/utils/observable/index.ts
@@ -1,23 +1,17 @@
 import { IObservable, IObserver } from './types';
 
 export class Observable implements IObservable {
-    private readonly observers: IObserver[] = [];
+    private readonly observers = new Set<IObserver>();
 
     public addObserver(observer: IObserver) {
-        this.observers.push(observer);
+        this.observers.add(observer);
     }
 
     public deleteObserver(observer: IObserver) {
-        const observerIndex = this.observers.indexOf(observer);
-        if (observerIndex < 0) {
-            return;
-        }
-        this.observers.splice(observerIndex, 1);
+        this.observers.delete(observer);
     }
 
     public notifyObservers() {
-        this.observers.forEach(observer => {
-            observer.update(this);
-        });
+        this.observers.forEach(observer => observer.update(this));
     }
 }
